# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeReconOps API - Cross-border financial reconciliation platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export interface UploadResult {
  batchId: string;
  rowCount: number;
  fileName: string;
  uploadType: string;
  errors: string[];
  warnings: string[];
}

export interface ReconciliationSettings {
  /** Fuzzy matching threshold (0-100) */
  fuzzyThreshold?: number;
  /** Date matching window in days */
  dateWindowDays?: number;
  /** FX variance tolerance percentage */
  fxVariancePct?: number;
  /** FX variance tolerance absolute CNY amount */
  fxVarianceAbs?: number;
}

export interface ReconcileRequest {
  bankBatchId: string;
  ledgerBatchId: string;
  settings?: ReconciliationSettings;
}

export interface ReconcileSummary {
  matched: number;
  fxVariance: number;
  amountMismatch: number;
  unmatchedBank: number;
  unmatchedLedger: number;
  possibleDuplicate: number;
  manualReview: number;
  totalUnresolvedAmount: string;
}

export type ReconciliationResultStatus =
  (typeof ReconciliationResultStatus)[keyof typeof ReconciliationResultStatus];

export const ReconciliationResultStatus = {
  MATCHED: "MATCHED",
  FX_VARIANCE: "FX_VARIANCE",
  AMOUNT_MISMATCH: "AMOUNT_MISMATCH",
  UNMATCHED_BANK: "UNMATCHED_BANK",
  UNMATCHED_LEDGER: "UNMATCHED_LEDGER",
  POSSIBLE_DUPLICATE: "POSSIBLE_DUPLICATE",
  MANUAL_REVIEW: "MANUAL_REVIEW",
} as const;

export type ReconciliationResultRiskLevel =
  (typeof ReconciliationResultRiskLevel)[keyof typeof ReconciliationResultRiskLevel];

export const ReconciliationResultRiskLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export interface ReconciliationResult {
  id: string;
  batchId: string;
  /** @nullable */
  bankDate?: string | null;
  /** @nullable */
  bankDescription?: string | null;
  /** @nullable */
  bankAmount?: string | null;
  /** @nullable */
  bankCurrency?: string | null;
  /** @nullable */
  bankType?: string | null;
  /** @nullable */
  bankReference?: string | null;
  /** @nullable */
  ledgerDate?: string | null;
  /** @nullable */
  ledgerVendorClient?: string | null;
  /** @nullable */
  ledgerAmount?: string | null;
  /** @nullable */
  ledgerCurrency?: string | null;
  /** @nullable */
  ledgerInvoiceId?: string | null;
  /** @nullable */
  ledgerCategory?: string | null;
  /** @nullable */
  exchangeRate?: string | null;
  /** @nullable */
  convertedAmount?: string | null;
  /** @nullable */
  differenceAmount?: string | null;
  /** @nullable */
  differencePct?: string | null;
  status: ReconciliationResultStatus;
  riskLevel: ReconciliationResultRiskLevel;
  reason: string;
  /** @nullable */
  fuzzyScore?: number | null;
}

export interface ReconcileResponse {
  batchId: string;
  totalResults: number;
  riskScore: number;
  summary: ReconcileSummary;
  results: ReconciliationResult[];
}

export interface RiskScoreBreakdown {
  unmatchedCount: number;
  amountMismatchCount: number;
  duplicateCount: number;
  criticalCount: number;
  fxVarianceCount: number;
  formula: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalBankTransactions: number;
  totalLedgerEntries: number;
  matched: number;
  fxVarianceCount: number;
  realDiscrepancyCount: number;
  manualReviewCount: number;
  possibleDuplicateCount: number;
  totalUnresolvedAmount: string;
  riskScore: number;
  riskScoreBreakdown: RiskScoreBreakdown;
  statusDistribution: StatusCount[];
  /** @nullable */
  lastReconciliationAt?: string | null;
  /** @nullable */
  activeBatchId?: string | null;
}

export interface ResultsPage {
  results: ReconciliationResult[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface UploadBatch {
  id: string;
  /** @nullable */
  bankFileName?: string | null;
  /** @nullable */
  ledgerFileName?: string | null;
  /** @nullable */
  bankRowCount?: number | null;
  /** @nullable */
  ledgerRowCount?: number | null;
  status: string;
  createdAt: string;
  /** @nullable */
  reconciledAt?: string | null;
  /** @nullable */
  riskScore?: number | null;
}

export interface FxRate {
  date: string;
  usdToCny: string;
  cnyToUsd: string;
  source: string;
}

export interface AiRequest {
  batchId: string;
}

export interface RiskExplanation {
  resultId: string;
  explanation: string;
  riskLevel: string;
}

export interface AiAuditSummary {
  batchId: string;
  summary: string;
  highRiskExplanations: RiskExplanation[];
  generatedAt: string;
  provider: string;
}

export type RecommendedActionPriority =
  (typeof RecommendedActionPriority)[keyof typeof RecommendedActionPriority];

export const RecommendedActionPriority = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface RecommendedAction {
  priority: RecommendedActionPriority;
  category: string;
  action: string;
  affectedItems: number;
}

export interface AiRecommendedActions {
  batchId: string;
  actions: RecommendedAction[];
  generatedAt: string;
  provider: string;
}

export interface AuditReport {
  id: string;
  batchId: string;
  markdownContent: string;
  generatedAt: string;
  provider: string;
}

export type UploadBankStatementBody = {
  file: Blob;
};

export type UploadLedgerBody = {
  file: Blob;
};

export type GetDashboardParams = {
  /**
   * @nullable
   */
  batchId?: string | null;
};

export type GetResultsParams = {
  /**
   * @nullable
   */
  batchId?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  riskLevel?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  page?: number;
  pageSize?: number;
};

export type GetLatestReportParams = {
  /**
   * @nullable
   */
  batchId?: string | null;
};