# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeReconOps API - Cross-border financial reconciliation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/latest:
    get:
      operationId: getLatestReport
      tags: [reports]
      summary: Get latest audit report
      description: Returns the latest generated markdown audit report
      parameters:
        - name: batchId
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Latest report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
        "404":
          description: No report found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate audit report
      description: Generates and stores an audit-ready markdown report for a batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRequest"
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
...
    AiRequest:
      type: object
      properties:
        batchId:
          type: string
      required:
        - batchId
...
    AuditReport:
      type: object
      properties:
        id:
          type: string
        batchId:
          type: string
        markdownContent:
          type: string
        generatedAt:
          type: string
        provider:
          type: string
      required:
        - id
        - batchId
        - markdownContent
        - generatedAt
        - provider